rename Areshkin as areshkin_pole_sum
