+++ dptb/negf/density.py
@@ -1,6 +1,6 @@
 import torch
 from dptb.negf.ozaki_res_cal import Ozaki_residues
-from dptb.negf.Areshkin import pole_maker
+from dptb.negf.areshkin_pole_sum import pole_maker
 import numpy as np
 from dptb.negf.utils import gauss_xw
 import logging
