+++ dptb/negf/hamiltonian.py
@@ -1,11 +1,11 @@
 from typing import List
 import torch
-from dptb.negf.Areshkin import pole_maker
+from dptb.negf.areshkin_pole_sum import pole_maker
 from dptb.negf.RGF import recursive_gf
 from dptb.negf.surface_green import selfEnergy
 from dptb.negf.utils import quad, gauss_xw
 from dptb.negf.ozaki_res_cal import Ozaki_residues
-from dptb.negf.Areshkin import pole_maker
+from dptb.negf.areshkin_pole_sum import pole_maker
 from ase.io import read
 from dptb.negf.poisson import density2Potential, getImg
 from dptb.negf.SCF import _SCF
